accessible/tests/mochitest/treeupdate/test_tooltip.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible XUL tooltip test">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../promisified-events.js" />

  <script type="application/javascript">
  <![CDATA[

    async function doTest() {
      let tooltip = document.getElementById("tooltip");

      testAccessibleTree("tooltip-container", { GROUPING: [
      ] });

      let shown = waitForEvent(EVENT_SHOW, tooltip);
      tooltip.openPopup();
      await shown;

      testAccessibleTree("tooltip-container",
        { GROUPING: [
          { TOOLTIP: [] },
          { STATICTEXT: [] },
        ] });

      let hidden = waitForEvent(EVENT_HIDE, tooltip);
      tooltip.hidePopup();
      await hidden;

      testAccessibleTree("tooltip-container", { GROUPING: [] });

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=1652211"
         title="Added anonymous tooltip to mochitest docs messes with text">
        Bug 1652211
      </a>
      <br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1" role="group" id="tooltip-container">
      <tooltip id="tooltip">
        <description class="tooltip-label" value="hello world"/>
      </tooltip>
    </vbox>
  </hbox>

</window>
